requirements.txt
Flask>=3.0.0
python-barcode>=0.15.1
Pillow>=10.0.0
gunicorn>=22.0.0
